examples/python/cloudwatch_grapher/requirements.txt
gevent
pyzmq
boto